Support user replacement of default font table
The methods that print characters and strings to the LED segments already support custom fonts to be used. However, unless all the optinal parameters are given at each call, the TM16xx display module will revert to the default font for subsequent display calls.

This commit allows the TM16xx display's default font to be replaced by a user-defined font table. The TM16xx class is able to setup it's default font at construction or the default font can be changed on the fly using the new setDefaultAlphaFont() and setDefaultNumberFont() methods. Internal member variables store pointers to the custom font tables in use, and the default font tables will be used if no custom font table is given.